backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose
sqlalchemy
pydantic
requests
python-dotenv